pyproject.toml
[project]
name = "sticky-brain"
version = "0.1.0"
description = "A local-first personal memory system — capture, organize, retrieve, and synthesize knowledge with zero friction"
requires-python = ">=3.11"
readme = "README.md"
license = "MIT"
keywords = ["knowledge-management", "second-brain", "mcp", "semantic-search", "personal-memory"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: News/Diary",
    "Topic :: Text Processing :: General",
]
dependencies = [
    "mcp[cli]>=1.0",
    "sentence-transformers>=3.0",
    "sqlite-vec>=0.1",
    "typer>=0.12",
    "textual>=0.80",
    "pydantic>=2.0",
    "python-ulid>=3.0",
    "httpx>=0.27",
    "tomli>=2.0; python_version<'3.11'",
    "tomli-w>=1.0",
    "numpy>=1.26",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
]

[project.urls]
Homepage = "https://github.com/nothans/sticky"
Repository = "https://github.com/nothans/sticky"

[project.scripts]
sticky = "sticky.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sticky"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
